Apply style changes to all the docs (#74)
Recently, some style rules were agreed. Several of these are rolled out
in this pull request including:
 * Making heading styles consistent with recommended style
 * Adding Fortran code style to several code blocks
* Change section references to remove the `section` prefix, as it is not
applied consistently
* Fixing a few files that had long lines or paragraphs with no
line-breaks - an 80-character limit was appled.

This change may conflict widely, so be good to get it reviewed quickly.
It shouldn't impact the mesh documentation that is currently being
written, as that is mostly new.