Add deploy target to MakeFile (#114)
Adding a deploy target to the Makefile to allow the output to be
deployed to the `public_html` directory on AZSpice this can then be
linked in PRs and viewed by reviewers. The target creates a sub
directory named after the git branch with the generated docs.

See: https://wwwspice/~andrew.coughtrie/make_deploy/html

Co-authored-by: Mike Hobson <26921912+mike-hobson@users.noreply.github.com>